# docs/swagger.yaml
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.ERC20Token:
    properties:
      address:
        description: Address 代币合约的地址
        type: string
      decimals:
        description: Decimals 代币的小数位数
        type: integer
      description:
        description: Description 对代币的描述
        type: string
      id:
        type: integer
      logoURI:
        description: LogoURI 代币的图标 URI
        type: string
      name:
        description: Name 代币的名称
        type: string
      symbol:
        description: Symbol 代币的简称
        type: string
    type: object
...
  schema.GetReactionByObjectIdResp:
    properties:
      reaction_summary:
        items:
          $ref: '#/definitions/schema.ReactionRespItem'
        type: array
    type: object
...
  schema.ListTokenReq:
    type: object
  schema.ListTokenResp:
    properties:
      list:
        items:
          $ref: '#/definitions/schema.ERC20Token'
        type: array
    type: object
...
  schema.ReactionRespItem:
    properties:
      count:
        description: Count is the number of users who reacted
        type: integer
      emoji:
        description: Emoji is the reaction emoji
        type: string
      is_active:
        description: IsActive is if current user has reacted
        type: boolean
      tooltip:
        description: Tooltip is the user's name who reacted
        type: string
    type: object
...
  schema.SignReq:
    properties:
      data:
        type: string
    required:
    - data
    type: object
  schema.SignResp:
    properties:
      r:
        type: string
      s:
        type: string
      v:
        type: string
    type: object
  schema.SignV3exReq:
    properties:
      amount:
        type: integer
      chain_id:
        type: integer
      date:
        type: string
      method:
        type: string
      to:
        type: string
    required:
    - amount
    - chain_id
    - date
    - method
    - to
    type: object
  schema.SignV3exResp:
    properties:
      sign_hash:
        type: string
    type: object
...
  schema.UpdateReactionReq:
    properties:
      emoji:
        enum:
        - heart
        - smile
        - frown
        type: string
      object_id:
        type: string
      reaction:
        enum:
        - activate
        - deactivate
        type: string
    required:
    - emoji
    - object_id
    - reaction
    type: object
...
info:
  contact: {}
paths:
...
  /answer/api/v1/meta/reaction:
    get:
      consumes:
      - application/json
      description: get reaction for an object
      parameters:
      - description: object_id
        in: query
        name: object_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetReactionByObjectIdResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get reaction
      tags:
      - Meta
    put:
      consumes:
      - application/json
      description: update reaction. if not exist, add one
      parameters:
      - description: reaction
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateReactionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: add or update reaction
      tags:
      - Meta
...
  /answer/api/v1/v3ex/checkin/sign:
    post:
      consumes:
      - application/json
      description: sign v3ex checkin data
      parameters:
      - description: SignV3exReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SignV3exReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SignV3exResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: sign v3ex checkin data
      tags:
      - V3ex
  /answer/api/v1/v3ex/sign:
    post:
      consumes:
      - application/json
      description: sign v3ex data
      parameters:
      - description: SignReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SignReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SignResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: sign v3ex data
      tags:
      - V3ex
  /answer/api/v1/v3ex/token:
    delete:
      consumes:
      - application/json
      description: delete v3ex support erc20 token
      parameters:
      - description: ERC20Token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ERC20Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: delete v3ex support erc20 token
      tags:
      - V3ex
    post:
      consumes:
      - application/json
      description: add v3ex support erc20 token
      parameters:
      - description: ERC20Token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ERC20Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: add v3ex support erc20 token
      tags:
      - V3ex
    put:
      consumes:
      - application/json
      description: update v3ex support erc20 token
      parameters:
      - description: ERC20Token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ERC20Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update v3ex support erc20 token
      tags:
      - V3ex
  /answer/api/v1/v3ex/tokens:
    get:
      consumes:
      - application/json
      description: list v3ex support erc20 token
      parameters:
      - description: ListTokenReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ListTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ListTokenResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: list v3ex support erc20 token
      tags:
      - V3ex
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"